Move /documentation handler to frontend/
